ZastepcaProcesowy/requirements.txt
Flask>=1.1
Flask-Session
mysql-connector-python~=8.0.5
Flask-WTF
WTForms=2.3.3
email-validator
